added flatbuffers schemas for structuring dataloaders
