+++ src/data-tools/converters/structure.schema.json
@@ -1,0 +1,184 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "definitions": {
+    "AtomicModel" : {
+      "type" : "string",
+      "enum": ["Point", "Sphere", "Gaussian"]
+    },
+    "DataClass" : {
+      "type" : "string",
+      "enum": ["Sim", "Synth", "Exp"]
+    },
+    "Domain" : {
+      "type" : "string",
+      "enum": ["Real", "Frequency"]
+    },
+    "DataType" : {
+      "type" : "string",
+      "enum": ["Image", "Filter"]
+    },
+    "P" : {
+      "type" : "object",
+      "properties" : {
+        "x" : {
+                "type" : "number"
+              },
+        "y" : {
+                "type" : "number"
+              },
+        "z" : {
+                "type" : "number"
+              }
+      },
+      "additionalProperties" : false
+    },
+    "Quaternion" : {
+      "type" : "object",
+      "properties" : {
+        "r" : {
+                "type" : "number"
+              },
+        "i" : {
+                "type" : "number"
+              },
+        "j" : {
+                "type" : "number"
+              },
+        "k" : {
+                "type" : "number"
+              }
+      },
+      "additionalProperties" : false
+    },
+    "Atom" : {
+      "type" : "object",
+      "properties" : {
+        "pos" : {
+                "$ref" : "#/definitions/P"
+              },
+        "mass" : {
+                "type" : "number"
+              },
+        "name" : {
+                "type" : "number"
+              },
+        "model" : {
+                "$ref" : "#/definitions/AtomicModel"
+              }
+      },
+      "additionalProperties" : false
+    },
+    "Parameters" : {
+      "type" : "object",
+      "properties" : {
+        "amplitude" : {
+                "type" : "number"
+              },
+        "defocus" : {
+                "type" : "number"
+              },
+        "b_factor" : {
+                "type" : "number"
+              },
+        "img_dims" : {
+                "type" : "number"
+              },
+        "num_pixels" : {
+                "type" : "number"
+              },
+        "pixel_width" : {
+                "type" : "number"
+              },
+        "sigma" : {
+                "type" : "number"
+              },
+        "elecwavel" : {
+                "type" : "number"
+              },
+        "snr" : {
+                "type" : "number"
+              },
+        "experiment_parameters" : {
+                "type" : "number"
+              },
+        "seed" : {
+                "type" : "number"
+              },
+        "structures" : {
+                "type" : "number"
+              },
+        "coordinates" : {
+                "type" : "number"
+              }
+      },
+      "additionalProperties" : false
+    },
+    "Image" : {
+      "type" : "object",
+      "properties" : {
+        "pixels" : {
+                "type" : "array", "items" : { "type" : "number" }
+              },
+        "parent_structure" : {
+                "type" : "number"
+              },
+        "orientation_id" : {
+                "type" : "number"
+              },
+        "data_class" : {
+                "$ref" : "#/definitions/DataClass"
+              },
+        "data_space" : {
+                "$ref" : "#/definitions/Domain"
+              }
+      },
+      "additionalProperties" : false
+    },
+    "Filter" : {
+      "type" : "object",
+      "properties" : {
+        "pixels" : {
+                "type" : "array", "items" : { "type" : "number" }
+              },
+        "filter_class" : {
+                "$ref" : "#/definitions/DataClass"
+              },
+        "filter_space" : {
+                "$ref" : "#/definitions/Domain"
+              }
+      },
+      "additionalProperties" : false
+    },
+    "Structure" : {
+      "type" : "object",
+      "properties" : {
+        "atoms" : {
+                "type" : "array", "items" : { "$ref" : "#/definitions/Atom" }
+              },
+        "orientations" : {
+                "type" : "array", "items" : { "$ref" : "#/definitions/Quaternion" }
+              }
+      },
+      "additionalProperties" : false
+    },
+    "DataSet" : {
+      "type" : "object",
+      "properties" : {
+        "structures" : {
+                "type" : "array", "items" : { "$ref" : "#/definitions/Structure" }
+              },
+        "images" : {
+                "type" : "array", "items" : { "$ref" : "#/definitions/Image" }
+              },
+        "filters" : {
+                "type" : "array", "items" : { "$ref" : "#/definitions/Filter" }
+              },
+        "params" : {
+                "type" : "array", "items" : { "$ref" : "#/definitions/Parameters" }
+              }
+      },
+      "additionalProperties" : false
+    }
+  },
+  "$ref" : "#/definitions/DataSet"
+}
