+++ src/data-tools/generators/dataloader.py
@@ -1,0 +1,28 @@
+#!/usr/bin/python3
+from filegroup import *
+
+
+class Dataloader:
+    """
+    An abstraction for loading data from files
+    """
+
+    def load_coords(filename):
+        """Loads (presumed) coordinates from a file
+
+        Args:
+            filename (_type_): _description_
+
+        Returns:
+            numpy array
+        """
+        coord_arr = np.load(f"{filename}")
+
+        if 3 not in coord_arr.shape:
+            print(
+                "Coordinate file does not contain a tuple of x,y,z coordinates, is it correct?"
+            )
+            return []
+
+        return coord_arr
+
