+++ src/data-tools/generators/filegroup.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python3
+from tqdm import tqdm
+import torch
+import numpy as np
